requirements.txt
Flask==3.0.3
openai==1.23.2
python-dotenv==1.0.1
